Change BluetoothGattService::GattErrorCode to an enum class.
Change BluetoothGattService::GattErrorCode to enum classes to avoid
conflicts in enum values, as detected by -Wshadow.

Bug: 1344231
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3906106
Commit-Queue: Lei Zhang <thestig@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1049227}

+++ device/bluetooth/bluetooth_remote_gatt_characteristic_unittest.cc
@@ -431,7 +431,8 @@ TEST_F(BluetoothRemoteGattCharacteristicTest,
       }),
       base::BindLambdaForTesting(
           [&](BluetoothGattService::GattErrorCode error_code) {
-            ADD_FAILURE() << "unexpected error: " << error_code;
+            ADD_FAILURE() << "unexpected error: "
+                          << static_cast<int>(error_code);
             loop.Quit();
           }));
   SimulateGattCharacteristicWrite(characteristic1_);
@@ -1021,7 +1022,8 @@ TEST_F(BluetoothRemoteGattCharacteristicTest, MAYBE_WriteRemoteCharacteristic) {
       }),
       base::BindLambdaForTesting(
           [&](BluetoothGattService::GattErrorCode error_code) {
-            ADD_FAILURE() << "unexpected error: " << error_code;
+            ADD_FAILURE() << "unexpected error: "
+                          << static_cast<int>(error_code);
             loop.Quit();
           }));
 
@@ -1149,7 +1151,7 @@ TEST_F(BluetoothRemoteGattCharacteristicTest,
       }),
       base::BindLambdaForTesting(
           [&](BluetoothGattService::GattErrorCode error_code) {
-            ADD_FAILURE() << "unexpected error" << error_code;
+            ADD_FAILURE() << "unexpected error" << static_cast<int>(error_code);
             loop1.Quit();
           }));
 
@@ -1168,7 +1170,7 @@ TEST_F(BluetoothRemoteGattCharacteristicTest,
       }),
       base::BindLambdaForTesting(
           [&](BluetoothGattService::GattErrorCode error_code) {
-            ADD_FAILURE() << "unexpected error" << error_code;
+            ADD_FAILURE() << "unexpected error" << static_cast<int>(error_code);
             loop2.Quit();
           }));
 
@@ -1308,7 +1310,7 @@ TEST_F(BluetoothRemoteGattCharacteristicTest,
       }),
       base::BindLambdaForTesting(
           [&](BluetoothGattService::GattErrorCode error_code) {
-            ADD_FAILURE() << "unexpected error" << error_code;
+            ADD_FAILURE() << "unexpected error" << static_cast<int>(error_code);
             loop1.Quit();
           }));
   if (!IsClassicWin()) {
@@ -1327,7 +1329,7 @@ TEST_F(BluetoothRemoteGattCharacteristicTest,
       }),
       base::BindLambdaForTesting(
           [&](BluetoothGattService::GattErrorCode error_code) {
-            ADD_FAILURE() << "unexpected error" << error_code;
+            ADD_FAILURE() << "unexpected error" << static_cast<int>(error_code);
             loop2.Quit();
           }));
   if (!IsClassicWin()) {
@@ -1492,7 +1494,8 @@ TEST_F(BluetoothRemoteGattCharacteristicTest,
             }),
             base::BindLambdaForTesting(
                 [&](BluetoothGattService::GattErrorCode error_code) {
-                  ADD_FAILURE() << "unexpected error: " << error_code;
+                  ADD_FAILURE()
+                      << "unexpected error: " << static_cast<int>(error_code);
                   loop.Quit();
                 }));
 
@@ -1500,7 +1503,8 @@ TEST_F(BluetoothRemoteGattCharacteristicTest,
       }),
       base::BindLambdaForTesting(
           [&](BluetoothGattService::GattErrorCode error_code) {
-            ADD_FAILURE() << "unexpected error: " << error_code;
+            ADD_FAILURE() << "unexpected error: "
+                          << static_cast<int>(error_code);
             loop.Quit();
           }));
 
@@ -1591,7 +1595,7 @@ TEST_F(BluetoothRemoteGattCharacteristicTest,
       [&](absl::optional<BluetoothGattService::GattErrorCode> error_code,
           const std::vector<uint8_t>& data) {
         ASSERT_FALSE(error_code.has_value())
-            << "unexpected error: " << error_code.value();
+            << "unexpected error: " << static_cast<int>(error_code.value());
         EXPECT_EQ(1, gatt_read_characteristic_attempts_);
         EXPECT_EQ(0, gatt_write_characteristic_attempts_);
         EXPECT_EQ(test_vector_1, data);
@@ -1607,7 +1611,8 @@ TEST_F(BluetoothRemoteGattCharacteristicTest,
             }),
             base::BindLambdaForTesting(
                 [&](BluetoothGattService::GattErrorCode error_code) {
-                  ADD_FAILURE() << "unexpected error: " << error_code;
+                  ADD_FAILURE()
+                      << "unexpected error: " << static_cast<int>(error_code);
                   loop.Quit();
                 }));
 
@@ -1721,7 +1726,8 @@ TEST_F(BluetoothRemoteGattCharacteristicTest,
       }),
       base::BindLambdaForTesting(
           [&](BluetoothGattService::GattErrorCode error_code) {
-            ADD_FAILURE() << "unexpected error: " << error_code;
+            ADD_FAILURE() << "unexpected error: "
+                          << static_cast<int>(error_code);
             loop.Quit();
           }));
 
@@ -1954,7 +1960,8 @@ TEST_F(BluetoothRemoteGattCharacteristicTest, MAYBE_WriteSynchronousError) {
       }),
       base::BindLambdaForTesting(
           [&](BluetoothGattService::GattErrorCode error_code) {
-            ADD_FAILURE() << "unexpected error: " << error_code;
+            ADD_FAILURE() << "unexpected error: "
+                          << static_cast<int>(error_code);
             loop2.Quit();
           }));
   SimulateGattCharacteristicWrite(characteristic1_);
@@ -2073,7 +2080,8 @@ TEST_F(BluetoothRemoteGattCharacteristicTest,
       }),
       base::BindLambdaForTesting(
           [&](BluetoothGattService::GattErrorCode error_code) {
-            ADD_FAILURE() << "unexpected error: " << error_code;
+            ADD_FAILURE() << "unexpected error: "
+                          << static_cast<int>(error_code);
             loop1.Quit();
           }));
   base::RunLoop loop2;
@@ -2176,7 +2184,8 @@ TEST_F(BluetoothRemoteGattCharacteristicTest,
       }),
       base::BindLambdaForTesting(
           [&](BluetoothGattService::GattErrorCode error_code) {
-            ADD_FAILURE() << "unexpected error: " << error_code;
+            ADD_FAILURE() << "unexpected error: "
+                          << static_cast<int>(error_code);
             loop1.Quit();
           }));
   characteristic1_->ReadRemoteCharacteristic(base::BindLambdaForTesting(
@@ -2437,7 +2446,8 @@ TEST_F(BluetoothRemoteGattCharacteristicTest,
       }),
       base::BindLambdaForTesting(
           [&](BluetoothGattService::GattErrorCode error_code) {
-            ADD_FAILURE() << "unexpected error: " << error_code;
+            ADD_FAILURE() << "unexpected error: "
+                          << static_cast<int>(error_code);
             loop.Quit();
           }));
